Refactor AbstractLayout / BasicForceLayout
