+++ src/main/java/com/powsybl/nad/layout/LayoutParameters.java
@@ -10,27 +10,7 @@ package com.powsybl.nad.layout;
  * @author Florian Dupuy <florian.dupuy at rte-france.com>
  */
 public class LayoutParameters {
-    private double edgesForkAperture = Math.toRadians(60);
-    private double edgesForkLength = 0.8;
     private boolean textNodesForceLayout = false;
-
-    public double getEdgesForkAperture() {
-        return edgesForkAperture;
-    }
-
-    public LayoutParameters setEdgesForkAperture(double edgesForkApertureDegrees) {
-        this.edgesForkAperture = Math.toRadians(edgesForkApertureDegrees);
-        return this;
-    }
-
-    public double getEdgesForkLength() {
-        return edgesForkLength;
-    }
-
-    public LayoutParameters setEdgesForkLength(double edgesForkLength) {
-        this.edgesForkLength = edgesForkLength;
-        return this;
-    }
 
     public boolean isTextNodesForceLayout() {
         return textNodesForceLayout;
