+++ src/main/java/com/powsybl/nad/svg/SvgWriter.java
@@ -84,6 +84,10 @@ public class SvgWriter {
     private void writeSvg(Graph graph, OutputStream svgOs) {
         Objects.requireNonNull(graph);
         Objects.requireNonNull(svgOs);
+
+        // Edge coordinates need to be computed first, based on svg parameters
+        computeBranchEdgesCoordinates(graph);
+
         try {
             XMLStreamWriter writer = XmlUtil.initializeWriter(true, INDENT, svgOs);
             addSvgRoot(graph, writer);
@@ -99,6 +103,75 @@ public class SvgWriter {
         } catch (XMLStreamException e) {
             throw new UncheckedXmlStreamException(e);
         }
+    }
+
+    private void computeBranchEdgesCoordinates(Graph graph) {
+        graph.getNonMultiBranchEdgesStream().forEach(edge -> computeSingleBranchEdgeCoordinates(graph, edge));
+        graph.getMultiBranchEdgesStream().forEach(edges -> computeMultiBranchEdgesCoordinates(graph, edges));
+        graph.getThreeWtEdgesStream().forEach(edge -> computeThreeWtEdgeCoordinates(graph.getNode1(edge), graph.getNode2(edge), edge));
+        graph.getTextEdgesMap().forEach((edge, nodes) -> computeTextEdgeLayoutCoordinates(nodes.getFirst(), nodes.getSecond(), edge));
+    }
+
+    private void computeTextEdgeLayoutCoordinates(Node node1, Node node2, TextEdge edge) {
+        Point point1 = new Point(node1.getX(), node1.getY());
+        Point point2 = new Point(node2.getX(), node2.getY());
+        edge.setPoints(point1, point2);
+    }
+
+    private void computeSingleBranchEdgeCoordinates(Graph graph, BranchEdge edge) {
+        Node node1 = graph.getNode1(edge);
+        Node node2 = graph.getNode2(edge);
+        Point point1 = new Point(node1.getX(), node1.getY());
+        Point point2 = new Point(node2.getX(), node2.getY());
+        Point middle = Point.createMiddlePoint(point1, point2);
+        edge.setPoints1(point1, middle);
+        edge.setPoints2(point2, middle);
+    }
+
+    private void computeMultiBranchEdgesCoordinates(Graph graph, Set<Edge> edges) {
+        Edge firstEdge = edges.iterator().next();
+        Node node1 = graph.getNode1(firstEdge);
+        Node node2 = graph.getNode2(firstEdge);
+        Point pointA = new Point(node1.getX(), node1.getY());
+        Point pointB = new Point(node2.getX(), node2.getY());
+
+        double dx = pointB.getX() - pointA.getX();
+        double dy = pointB.getY() - pointA.getY();
+        double angle = Math.atan2(dy, dx);
+
+        int nbForks = edges.size();
+        double forkAperture = svgParameters.getEdgesForkAperture();
+        double forkLength = svgParameters.getEdgesForkLength();
+        double angleStep = forkAperture / (nbForks - 1);
+
+        int i = 0;
+        for (Edge edge : edges) {
+            if (!(edge instanceof BranchEdge)) {
+                continue;
+            }
+            BranchEdge branchEdge = (BranchEdge) edge;
+            if (2 * i + 1 == nbForks) { // in the middle, hence alpha = 0
+                computeSingleBranchEdgeCoordinates(graph, branchEdge);
+            } else {
+                double alpha = -forkAperture / 2 + i * angleStep;
+                double angleFork1 = angle - alpha;
+                double angleFork2 = angle + Math.PI + alpha;
+                Point fork1 = pointA.shift(forkLength * Math.cos(angleFork1), forkLength * Math.sin(angleFork1));
+                Point fork2 = pointB.shift(forkLength * Math.cos(angleFork2), forkLength * Math.sin(angleFork2));
+
+                Point middle = Point.createMiddlePoint(fork1, fork2);
+                BranchEdge.Side side = graph.getNode1(edge) == node1 ? BranchEdge.Side.ONE : BranchEdge.Side.TWO;
+                branchEdge.setPoints(side, pointA, fork1, middle);
+                branchEdge.setPoints(side.getOpposite(), pointB, fork2, middle);
+            }
+            i++;
+        }
+    }
+
+    private void computeThreeWtEdgeCoordinates(Node node1, Node node2, ThreeWtEdge edge) {
+        Point point1 = new Point(node1.getX(), node1.getY());
+        Point point2 = new Point(node2.getX(), node2.getY());
+        edge.setPoints(point1, point2);
     }
 
     private void drawBranchEdges(Graph graph, XMLStreamWriter writer) throws XMLStreamException {
